Got a unit test working, got the pipeline kind of refactored to output frames from the vis function insted of a visual file
